Crate trusted_database_benchmark

Trusted Database Benchmark.

This benchmark measures average gRPC request time to a database that is either:

A part of the Trusted Database Oak application

This version uses compiled Wasm modules and Oak IFC which both account for a nontrivial performance difference

Implemented in oak::OakApplication

A tonic gRPC server without using Oak

Implemented in native::NativeApplication Benchmark pseudo-randomly (but deterministically) generates a database and sends it both to Oak and native application. Then it sends REQUEST_NUMBER of requests to each application and computes the average response time. It also computes time needed to start Oak application (which includes XML database deserialization).

It’s important to note that Wasm module for Trusted Database application is being compiled with --release.

Invoke with:

cargo run --release --manifest-path=experimental/benchmark/Cargo.toml -- --database-size 100

Where --database-size specifies the size (number of entries) of the generated database.

Modules

application

database

proto

Structs

Opt

Constants

NATIVE_APPLICATION_GRPC_PORT

REQUEST_NUMBER

Functions

main

measure_request_time

Measures average request time in milliseconds.