Trait trusted_database_benchmark::proto::trusted_database_server::TrustedDatabase

pub trait TrustedDatabase: Send + Sync + 'static {
    fn get_point_of_interest<'life0, 'async_trait>(
        &'life0 self, 
        request: Request<GetPointOfInterestRequest>
    ) -> Pin<Box<dyn Future<Output = Result<Response<GetPointOfInterestResponse>, Status>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

    fn list_points_of_interest<'life0, 'async_trait>(
        &'life0 self, 
        request: Request<ListPointsOfInterestRequest>
    ) -> Pin<Box<dyn Future<Output = Result<Response<ListPointsOfInterestResponse>, Status>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
}

Generated trait containing gRPC methods that should be implemented for use with TrustedDatabaseServer.

Required methods

fn get_point_of_interest<'life0, 'async_trait>(
&'life0 self,
request: Request<GetPointOfInterestRequest>
) -> Pin<Box<dyn Future<Output = Result<Response<GetPointOfInterestResponse>, Status>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Find Point Of Interest based on ID.

fn list_points_of_interest<'life0, 'async_trait>(
&'life0 self,
request: Request<ListPointsOfInterestRequest>
) -> Pin<Box<dyn Future<Output = Result<Response<ListPointsOfInterestResponse>, Status>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Find the nearest Point Of Interest based on linear search in the database.